ConvexiPy/requirements.txt
opencv-contrib-python==4.4.0.44
pathlib
#replace PIL with Pillow please
Pillow==7.2.0
scipy
numpy
matplotlib